+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Changelog for NeoFS Node
 - IR, CLI and SN ignore signatures of object GET/HEAD responses (#3406)
 - IR, CLI and SN verify object checksums from GET/HEAD responses (except SN proxy case) (#3406)
 - SN does not sign object GET/HEAD responses to requests with API version > v2.17 (#3406)
+- SN now verifies object header from proxy GET/HEAD responses against requested ID (#3406)
 
 ### Removed